backend/requirements.txt
Flask
langchain
langchain_chroma
langchain_community
requests
PyPDF2
python-docx
werkzeug
uuid
logging
